preview button functionality and styling put in testmenu component for react, redux, and hooks. switching to master to pull down latest version.
